backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
bcrypt==4.1.3
beautifulsoup4==4.14.2
bidict==0.23.1
black==26.1.0
bleach==6.3.0
boto3==1.42.58
botocore==1.42.58
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.110.1
fastjsonschema==2.21.2
fastuuid==0.14.0
filelock==3.25.0
firebase_admin==7.2.0
flake8==7.3.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-genai==1.64.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==8.0.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jq==1.11.0
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_git==0.51.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lark==1.3.1
librt==0.8.1
limits==5.8.0
litellm==1.80.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
motor==3.3.1
pymongo==4.5.0

# Authentication & Security
python-jose==3.5.0
bcrypt==4.1.3
passlib==1.7.4
pyotp==2.9.0
cryptography==46.0.5

# Data Validation
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
email-validator==2.3.0
python-dotenv==1.2.1

# HTTP & Networking
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
aiohttp==3.13.3
websockets==15.0.1

# WebSocket / Real-time
python-socketio==5.16.1
python-engineio==4.13.1
simple-websocket==1.1.0

# Email
aiosmtplib==5.1.0
sendgrid==6.12.5
python-http-client==3.3.7

# Firebase
firebase_admin==7.2.0
google-cloud-firestore==2.25.0
google-cloud-storage==3.9.0
google-auth==2.49.0.dev0
google-api-core==2.30.0

# Redis (optional - circuit breaker handles unavailability)
redis==7.2.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
Send2Trash==1.8.3
sendgrid==6.12.5
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
limits==5.8.0

# Payments
stripe==14.3.0

# Monitoring
sentry-sdk[fastapi]

# Logging
python-json-logger==4.0.0

# Utilities
python-dateutil==2.9.0.post0
PyYAML==6.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
typing_extensions==4.15.0
certifi==2026.2.25
urllib3==2.6.3
idna==3.11
sniffio==1.3.1
anyio==4.12.1
click==8.3.1
h11==0.16.0
dnspython==2.8.0
six==1.17.0
attrs==25.4.0
charset-normalizer==3.4.4
frozenlist==1.8.0
multidict==6.7.1
yarl==1.22.0
aiosignal==1.4.0
bidict==0.23.1
propcache==0.4.1
annotated-types==0.7.0
cffi==2.0.0
pycparser==3.0
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
proto-plus==1.27.1
protobuf==5.29.6
grpcio==1.78.1
grpcio-status==1.71.2
googleapis-common-protos==1.72.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
google-api-python-client==2.191.0
google-auth-httplib2==0.3.0
httplib2==0.31.2
uritemplate==4.2.0
pyparsing==3.3.2
oauthlib==3.3.1
Deprecated==1.3.1
wrapt==2.1.1
packaging==26.0
wsproto==1.3.2
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
Werkzeug==3.1.6
